requirements.txt
langchain==0.3.25
qdrant-client==1.14.2
sentence-transformers==4.1.0
langchain-openai
python-dotenv
tqdm>=4.66.0
langchain-community
langchain_huggingface
langchain_qdrant